Flatten out `fluvio hub`
Resolves #3187
